+++ lib/volgo_base/src/platform_repo.ml
@@ -19,22 +19,58 @@
 (*  <http://www.gnu.org/licenses/> and <https://spdx.org>, respectively.       *)
 (*******************************************************************************)
 
+module Vcs_kind = struct
+  module T0 = struct
+    type t = Vcs.Platform_repo.Vcs_kind.t =
+      | Git
+      | Hg
+    [@@deriving hash]
+  end
+
+  include (
+    Vcs.Platform_repo.Vcs_kind :
+      module type of Vcs.Platform_repo.Vcs_kind with type t := T0.t)
+
+  include T0
+end
+
 module Protocol = struct
   module T0 = struct
-    type t = Vcs.Url.Protocol.t =
+    type t = Vcs.Platform_repo.Protocol.t =
       | Ssh
       | Https
     [@@deriving hash]
   end
 
-  include (Vcs.Url.Protocol : module type of Vcs.Url.Protocol with type t := T0.t)
+  include (
+    Vcs.Platform_repo.Protocol :
+      module type of Vcs.Platform_repo.Protocol with type t := T0.t)
+
+  include T0
+end
+
+module Url = struct
+  module T0 = struct
+    type t = Vcs.Platform_repo.Url.t =
+      { platform : Platform.t
+      ; vcs_kind : Vcs_kind.t
+      ; user_handle : User_handle.t
+      ; repo_name : Repo_name.t
+      ; protocol : Protocol.t
+      }
+    [@@deriving hash]
+  end
+
+  include (
+    Vcs.Platform_repo.Url : module type of Vcs.Platform_repo.Url with type t := T0.t)
+
   include T0
 end
 
 module T0 = struct
-  type t = Vcs.Url.t =
+  type t = Vcs.Platform_repo.t =
     { platform : Platform.t
-    ; protocol : Protocol.t
+    ; vcs_kind : Vcs_kind.t
     ; user_handle : User_handle.t
     ; repo_name : Repo_name.t
     }
@@ -42,6 +78,11 @@ module T0 = struct
 end
 
 include (
-  Vcs.Url : module type of Vcs.Url with type t := T0.t and module Protocol := Protocol)
+  Vcs.Platform_repo :
+    module type of Vcs.Platform_repo
+    with type t := T0.t
+     and module Protocol := Protocol
+     and module Url := Url
+     and module Vcs_kind := Vcs_kind)
 
 include T0